+++ TeslaLogger/DBHelper.cs
@@ -6,7 +6,6 @@ using System.Globalization;
 using System.Net;
 using System.Runtime.Caching;
 using System.Text;
-using System.Threading;
 using System.Threading.Tasks;
 using System.Web.Script.Serialization;
 
@@ -25,7 +24,6 @@ namespace TeslaLogger
 
         public static void EnableMothership()
         {
-            GetMothershipCommandsFromDB();
             mothershipEnabled = true;
         }
 
@@ -487,14 +485,7 @@ namespace TeslaLogger
             currentJSON.current_speed = 0;
             currentJSON.current_power = 0;
 
-            Thread TripElevationUpdater = new Thread(() =>
-            {
-                UpdateTripElevation(StartPos, MaxPosId);
-            })
-            {
-                Priority = ThreadPriority.BelowNormal
-            };
-            TripElevationUpdater.Start();
+            UpdateTripElevation(StartPos, MaxPosId);
         }
 
         public static void UpdateTripElevation(int startPos, int maxPosId)
@@ -528,6 +519,8 @@ namespace TeslaLogger
                     string sql = null;
                     try
                     {
+                        System.Threading.Thread.Sleep(1);
+
                         double latitude = (double)dr[1];
                         double longitude = (double)dr[2];
 
@@ -717,6 +710,8 @@ namespace TeslaLogger
         {
             try
             {
+                System.Threading.Thread.Sleep(5);
+
                 using (MySqlConnection con = new MySqlConnection(DBConnectionstring))
                 {
                     con.Open();
@@ -1616,6 +1611,8 @@ namespace TeslaLogger
         {
             try
             {
+                System.Threading.Thread.Sleep(5);
+
                 using (MySqlConnection con = new MySqlConnection(DBConnectionstring))
                 {
                     con.Open();
