+++ TeslaLogger/Geofence.cs
@@ -17,9 +17,7 @@ namespace TeslaLogger
         {
             OpenChargePort,
             HighFrequencyLogging,
-            EnableSentryMode,
-            SetChargeLimit,
-            ClimateOff
+            EnableSentryMode
         }
 
         public string name;
@@ -237,23 +235,6 @@ namespace TeslaLogger
                     {
                         Logfile.ExceptionWriter(ex, line);
                     }
-                    foreach (Address addr in list)
-                    {
-                        bool keepAddr = true;
-                        foreach (string localName in uniqueNameList)
-                        {
-                            if (addr != null && addr.name != null && localName != null && localName.Equals(addr.name))
-                            {
-                                Logfile.Log("replace " + addr.name + " with value(s) from " + filename);
-                                keepAddr = false;
-                                break;
-                            }
-                        }
-                        if (keepAddr)
-                        {
-                            localList.Add(addr);
-                        }
-                    }
                 }
             }
         }
@@ -317,14 +298,6 @@ namespace TeslaLogger
                 {
                     _addr.IsWork = true;
                 }
-                else if (flag.StartsWith("scl"))
-                {
-                    SpecialFlag_SCL(_addr, flag);
-                }
-                else if (flag.StartsWith("cof"))
-                {
-                    SpecialFlag_COF(_addr, flag);
-                }
             }
         }
 
@@ -334,42 +307,12 @@ namespace TeslaLogger
             Match m = Regex.Match(_flag, pattern);
             if (m.Success && m.Groups.Count == 3 && m.Groups[1].Captures.Count == 1 && m.Groups[2].Captures.Count == 1)
             {
-                _addr.specialFlags.Add(Address.SpecialFlags.EnableSentryMode, m.Groups[1].Captures[0].ToString() + "->" + m.Groups[2].Captures[0].ToString());
+                _addr.specialFlags.Add(Address.SpecialFlags.EnableSentryMode, m.Groups[0].Captures[1].ToString() + "->" + m.Groups[0].Captures[2].ToString());
             }
             else
             {
                 // default
                 _addr.specialFlags.Add(Address.SpecialFlags.EnableSentryMode, "RND->P");
-            }
-        }
-
-        private static void SpecialFlag_COF(Address _addr, string _flag)
-        {
-            string pattern = "cof:([PRND]+)->([PRND]+)";
-            Match m = Regex.Match(_flag, pattern);
-            if (m.Success && m.Groups.Count == 3 && m.Groups[1].Captures.Count == 1 && m.Groups[2].Captures.Count == 1)
-            {
-                _addr.specialFlags.Add(Address.SpecialFlags.ClimateOff, m.Groups[1].Captures[0].ToString() + "->" + m.Groups[2].Captures[0].ToString());
-            }
-            else
-            {
-                // default
-                _addr.specialFlags.Add(Address.SpecialFlags.ClimateOff, "RND->P");
-            }
-        }
-
-        private static void SpecialFlag_SCL(Address _addr, string _flag)
-        {
-            string pattern = "scl:([0-9]+)";
-            Match m = Regex.Match(_flag, pattern);
-            if (m.Success && m.Groups.Count == 2 && m.Groups[1].Captures.Count == 1)
-            {
-                _addr.specialFlags.Add(Address.SpecialFlags.SetChargeLimit, m.Groups[1].Captures[0].ToString());
-            }
-            else
-            {
-                // default
-                _addr.specialFlags.Add(Address.SpecialFlags.SetChargeLimit, "80");
             }
         }
