+++ TeslaLogger/Geofence.md
@@ -25,7 +25,7 @@ geofence-private.csv contains data records. One line is one record. Records cons
 4. Radius. Optional. Integer. Value in meters. Default: 50
 5. special flags. Optional. List of special flags and parameters, each flag starts with + Default: nothing
 
-Example 1
+Example
 
 Supercharger DE-Blankenfelde-Mahlow,52.308921402353334,13.444712162017824,20,+ocp:R->P+hfl:5m
 * Name: Supercharger DE-Blankenfelde-Mahlow
@@ -33,14 +33,6 @@ Supercharger DE-Blankenfelde-Mahlow,52.308921402353334,13.444712162017824,20,+oc
 * Longitude: 13.444712162017824
 * Radius: 20
 * special flags: +ocp:R->P+hfl:5m
-
-Example 2
-My Home,52.39408202469785,13.542231917381288,,+scl:75
-* Name: My Home
-* Latitude: 52.39408202469785
-* Longitude: 13.542231917381288
-* Radius: empty, default of 50 is applied
-* special flags: +scl:75
 
 ## replacing TeslaLogger POIs
 
@@ -64,7 +56,7 @@ Configuration parameters: from->to
 
 One or more shift states P N R D are valid for from and to.
 
-Example: +ocp:R->P when shifting from R to P the command Open charge Port is sent to your car
+Example: R->P when shifting from R to P the command Open charge Port is sent to your car
 
 Default: RND->P
 
@@ -72,8 +64,7 @@ Default: RND->P
 
 +hfl
 
-This flag is executed when charing starts. It will enable high frequency logging which tries to poll
-the Tesla API as often as possible.
+This flag is executed during charging.
 
 Configuration parameters: duration
 
@@ -81,91 +72,10 @@ or
 
 Configuration parameters: count
 
-Example: +hfl:5m (log as fast as possible for 5 minutes)
+Example: 5m (log as fast as possible for 5 minutes)
 
 Valid configuration parameters: 
 
 count: integer
 
 duration: integer + one character (s: seconds, m: minutes, h: hours, d: days)
-
-Default: count 100
-
-# Enable Sentry Mode
-
-+esm
-
-Turn Sentry Mode on
-
-Warning: in case of network issues or other communication problems, Sentry Mode will not be turned on!
-
-This flag is executed when the shift state of your car changes, eg. when shifting from D to R.
-
-Configuration parameters: from->to
-
-One or more shift states P N R D are valid for from and to.
-
-Example: +esm:R->P when shifting from R to P the command Enable Sentry Mode is sent to your car
-
-Default: RND->P
-
-# Home Address
-
-+home
-
-Marks this POI as your home address.
-
-TODO: useful application
-
-Home cannot be work.
-
-# Work Address
-
-+work
-
-Marks this POI as you work address.
-
-TODO: useful application
-
-Work cannot be home.
-
-# Set Charge Limit
-
-+scl
-
-This flag is executed when charing starts. It will set the charge limit.
-
-Warning: in case of network issues or other communication problems, charge limit will not be set!
-
-This setting will always be set when TelsaLogger changes into state charging. You can change the charge
-limit again in the car and/or in the app after TeslaLogger has set the charge limit. Your manual
-override will not be changed, except in case TeslaLogger restarts (then TL will set TL's charge limit 
-again).
-
-Configuration parameters: limit
-
-Example: +scl:75 (set charge limit to 75%)
-
-Valid configuration parameters: 
-
-limit: integer
-
-Default: limit 80
-
-# Turn HVAC off
-
-+cof
-
-This will turn HVAC off.
-
-Warning: in case of network issues or other communication problems, HVAC will not be turned off!
-
-This flag is executed when the shift state of your car changes, eg. when shifting from D to R.
-
-Configuration parameters: from->to
-
-One or more shift states P N R D are valid for from and to.
-
-Example: +cof:R->P when shifting from R to P the command HVAC off is sent to your car
-
-Default: RND->P
