+++ TeslaLogger/bin/language-en.txt
@@ -192,5 +192,3 @@ Automatische Updates=Automatic updates
 
 TextShare=<h1>Share charge curves / degradation curves anonymously</h1>We would like to create an anonymous database to compare degradation curves and charge curves. This allows you to compare your vehicle with others at any time and to detect any unusual deviation at an early stage. This can also be used later to prove during the sales talk that your battery is still fully efficient. In the negative case, you can more easily prove a reduced performance against Tesla Motors.<br><br>Would you like to share your fast charging curves and degradation curves with other Tesla drivers completely anonymously? This decision can be changed at any time in the settings screen.
 Daten anonym teilen=Share data anonymously
-TextSuspendTeslalogger=Teslalogger has been stopped! Data will not be logged!
-Resume Teslalogger=Resume Teslalogger
