+++ TeslaLogger/bin/language-nl.txt
@@ -192,5 +192,4 @@ Automatische Updates=Automatiske oppdateringer
 
 TextShare=<h1>Share charge curves / degradation curves anonymously</h1>We would like to create an anonymous database to compare degradation curves and charge curves. This allows you to compare your vehicle with others at any time and to detect any unusual deviation at an early stage. This can also be used later to prove during the sales talk that your battery is still fully efficient. In the negative case, you can more easily prove a reduced performance against Tesla Motors.<br><br>Would you like to share your fast charging curves and degradation curves with other Tesla drivers completely anonymously? This decision can be changed at any time in the settings screen.
 Daten anonym teilen=Data anoniem delen
-TextSuspendTeslalogger=Teslalogger is gestopt! De gegevens worden niet gelogd!
-Resume Teslalogger=Start Teslalogger
+Daten anonym teilen=Data anoniem delen
