+++ TeslaLogger/bin/language-no.txt
@@ -161,5 +161,3 @@ Automatische Updates=Automatiske oppdateringer
 
 TextShare=<h1> Del ladekurver / nedbrytningskurver anonymt </h1> Vi ønsker å sette opp en anonym database som du kan sammenligne degraderingskurver og ladekurver med. Så du kan sammenligne kjøretøyet ditt med andre når som helst og identifisere et uvanlig avvik tidlig. Dette kan også brukes senere for å kunne bevise på salgshøyden at batteriet fremdeles er fullt effektivt. I negativt tilfelle er det lettere å bevise en redusert ytelse sammenlignet med Tesla Motors. <br> <br> Vil du dele hurtigladekurver og degraderingskurver anonymt med andre Tesla-drivere? Denne avgjørelsen kan endres når som helst i innstillingene.
 Daten anonym teilen=Del data anonymt
-TextSuspendTeslalogger=Teslalogger ble stoppet! Data logges ikke!
-Resume Teslalogger=Fortsett Teslalogger
