+++ TeslaLogger/www/admin/current_json.php
@@ -1,4 +1,3 @@
 <?PHP
-  header('Content-Type: application/json');
   echo(file_get_contents("/etc/teslalogger/current_json.txt"));
-?>
+  ?>
