+++ TeslaLogger/www/admin/db.php
@@ -4,11 +4,7 @@ $mysqluser="root";
 $mysqlpwd="teslalogger"; 
 $mysqldb="teslalogger"; 
 
-$con=new mysqli($mysqlhost, $mysqluser, $mysqlpwd, $mysqldb);
-/* check connection */
+$connection=mysql_connect($mysqlhost, $mysqluser, $mysqlpwd) or die ("Verbindungsversuch fehlgeschlagen: " . mysql_error());
+mysql_select_db($mysqldb, $connection) or die("Konnte die Datenbank nicht waehlen.");
 
-if (mysqli_connect_errno()) {
-    printf("Connect failed: %s\n", mysqli_connect_error());
-    exit();
-}
 ?>
