+++ TeslaLogger/www/wakeup.php
@@ -1,12 +1,5 @@
 <?PHP
-	require("admin/tools.php");
-
-	$prefix = "/etc/teslalogger/";
-    if (isDocker())
-        $prefix = "/tmp/";
-
-	unlink($prefix."cmd_gosleep.txt");
-	$filename = $prefix."wakeupteslalogger.txt";
+	$filename = "/etc/teslalogger/wakeupteslalogger.txt";
 	
 	$ret =	file_put_contents ($filename,".");
 	chmod ($filename, 0666);
